// frontend/src/api/generated/models/postApiRoomsBody.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export type PostApiRoomsBody = {
  Name?: string;
  Capacity?: number;
  LocationId?: number;
  Purpose?: string;
  image?: Blob;
};